backend/requirement.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
asttokens==2.4.1
catboost==1.2.7
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
colorlog==6.9.0
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.9
decorator==5.1.1
distro==1.9.0
dj-database-url==2.3.0
Django==5.1.3
django-cors-headers==4.6.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
executing==2.1.0
fonttools==4.55.0
graphviz==0.20.3
greenlet==3.1.1
groq==0.13.0
gTTS==2.5.4
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
ipykernel==6.29.5
ipython==8.30.0
jedi==0.19.2
joblib==1.4.2
kiwisolver==1.4.7
lightgbm==4.5.0
lxml==5.3.0
Mako==1.3.6
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
moviepy==2.1.1
nest-asyncio==1.6.0
numpy==1.26.4
optuna==4.1.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==10.4.0
platformdirs==4.3.6
plotly==5.24.1
plum-dispatch==1.7.4
proglog==0.1.10
prompt_toolkit==3.0.48
psutil==6.1.0
psycopg2-binary==2.9.10
pure_eval==0.2.3
pydantic==2.10.2
pydantic_core==2.27.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-pptx==1.0.2
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
requests==2.32.3
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
spire==0.4.2
Spire.Presentation==9.11.2
SQLAlchemy==2.0.36
sqlparse==0.5.2
stack-data==0.6.3
tenacity==9.0.0
threadpoolctl==3.5.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
xgboost==2.1.2
XlsxWriter==3.2.0